requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
openai>=1.40.0
pydantic>=2.5.0
PyPDF2>=3.0.0
python-docx>=1.0.0
